% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustytools.R
\name{align}
\alias{align}
\title{Aligns two strings using the specified alignment type.}
\usage{
align(s1, s2, atype = c("local", "global", "semi-global"), verbose = F)
}
\arguments{
\item{s1}{A character string representing the first sequence to align.}

\item{s2}{A character string representing the second sequence to align.}

\item{atype}{A character string specifying the type of alignment.
Possible values are "local", "global", or "semi-global". Default is "local".}

\item{verbose}{A logical value indicating whether to print detailed output. Default is FALSE.}
}
\value{
The alignment score.
}
\description{
This function aligns two strings using a Rust backend and returns the alignment score.
}
\examples{
\dontrun{
align("AGCT", "AGC", "global", TRUE)
align("AGCT", "AGC")
}
}
\references{
This documentation was written by ChatGPT v4o - OpenAI, conversation with the author, 6-5-2024.
}
