% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{dgC_from_components}
\alias{dgC_from_components}
\title{Build dgCMatrix from MAGIC Rust components}
\usage{
dgC_from_components(comps)
}
\arguments{
\item{comps}{list with elements \code{p}, \code{i}, \code{x}, \code{Dim}}
}
\value{
a \link[Matrix]{dgCMatrix}
}
\description{
Build dgCMatrix from MAGIC Rust components
}
