% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pcha}
\alias{pcha}
\title{Call PCHA from R}
\usage{
pcha(input_mat, noc)
}
\arguments{
\item{input_mat}{numeric matrix (variables in rows, samples in columns)}

\item{noc}{integer, number of archetypes/components}
}
\value{
list of matrices
}
\description{
Call PCHA from R
}
