% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{rustytools-package}
\alias{rustytools}
\alias{rustytools-package}
\title{rustytools: This package brings several rust tools to the R environment}
\description{
not that complicated
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scfurl.github.io/rustytools/}
}

}
\author{
\strong{Maintainer}: Scott Furlan \email{scottfurlan@gmail.com} (\href{https://orcid.org/0000-0003-3007-9840}{ORCID})

}
