% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustytools.R
\name{get_consensus}
\alias{get_consensus}
\title{Get Consensus Sequences from a FASTA File}
\usage{
get_consensus(fasta, cores = 1, genome = "hg38", test_with_n = NULL)
}
\arguments{
\item{fasta}{A string specifying the path to the input FASTA file.}

\item{cores}{An integer specifying the number of cores to use for parallel processing. Default is 1.}

\item{genome}{A string specifying the genome build. Default is "hg38".}

\item{test_with_n}{An optional integer specifying the number of sequences to process for testing. If NULL, all sequences are processed. Default is NULL.}
}
\value{
A GRangesList containing the consensus sequences with their coordinates and sequence information.
}
\description{
This function reads a DNA FASTA file, processes the sequences, and returns consensus sequences
as a GRangesList. The function can optionally use multiple cores for parallel processing.
}
\examples{
\dontrun{
consensus <- get_consensus("path/to/fasta/file.fasta", cores=2, genome="hg19")
}
}
\references{
This documentation was written by ChatGPT v4o - OpenAI, conversation with the author, 6-5-2024.
}
