% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustytools.R
\name{get_consensus}
\alias{get_consensus}
\title{Get consensus}
\usage{
get_consensus(str)
}
\description{
It will take a string and print the coordinates of every repeated character substring made up of either "ACGT".
}
